Move to generic table instead of inventoryTable
